Refactor the IP range calculation logic to fix the issue with except
This patch will fix the calculation issue for IP block with except
and move the IP related functions to a separate file.

Testing done:

create the following NetworkPolicy:

```
apiVersion: networking.k8s.io/v1
kind: NetworkPolicy
metadata:
  name: np1
spec:
  podSelector:
    matchLabels:
      app: coffee
  ingress:
  - from:
    - ipBlock:
        cidr: 40.0.0.0/8
        except:
        - 40.100.100.11/32
        - 40.100.100.15/32
        - 40.100.100.31/32
        - 40.100.100.32/32
    ports:
    - protocol: TCP
      port: 6000
    - protocol: UDP
      port: 6001
      endPort: 6010
```
and check the IP addresses in NSX are expected:
```
40.100.100.12-40.100.100.14
40.100.100.33-40.255.255.255
40.0.0.0-40.100.100.10
40.100.100.16-40.100.100.30
```
